Changes to support LPC11Uxx
This adds a linker script for the LPC11U24, and adds  CMSIS files
for the LPC11Uxx family. It changes the Makefile to use them and
modifies main to cycle the RGB LED on a Robotiky "brain" board.

It also removes the lpc11xx files, which apparently hold nothing in
common with the LPC11Uxx family of chips.