lpc_checksum: Fix alignment bug
How stupid are you Brian? Can't even write a simple 'cp' implementation.

Allow me to elaborate...
When writing the checksummed image to a new file, the copy of all the
data after the checksum was done word-wise (4 byte blocks). This means
that if the image was not a word-aligned length, we simply truncate
either 1, 2 or 3 bytes from the end of the image.

This manifests itself as corruption of the end of the last section in
the image (probably .data), which shows up as corrupted global
variables.
It's hard to track down - because it manifests itself in weird ways
(until you figure out it's related to the length of the image:)
 - Sometimes the value looks OK if you got lucky and flash already
   had what you want in it.
 - If all your .rodata and .data are word-aligned, no problem
 - If your last .rodata section has a non-word-aligned variable,
   you hit the bug. This makes it depend on link-order.
 - If you only check the binary *before* checksumming it (like I was
   doing), you spend 2 nights pulling your hair out.