usb_cdc: Add option to remove Call Management decriptor
I think there's a bug in the ROM driver, which makes enumeration
unreliable when the descriptors are larger than 64 bytes (one packet).
Add a #define to allow us to remove the Call Management Functional
Descriptor which brings the total size down to 62 bytes and makes
enumeration work reliably.

This might be a violation of the protocol, and might not actually work
that well on some platforms, so needs testing and using with care.
Works fine on eva and jon though, so enable by default.

Reported on NXP forum:
https://www.lpcware.com/content/forum/usbd-get-device-configuration-descriptor-truncated-packets